+++ discord-gateway/include/celery.h
@@ -1,5 +1,0 @@
-#include <nlohmann/json.hpp>
-
-namespace celery {
-void send_guild_member_add(nlohmann::json& op);
-}