+++ CHANGELOG.md
@@ -22,3 +22,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed `ddtrace` importing and setting of user ID in span within `@app.before_request`
+- Removed `discord-gateway` from monorepo